@@ -14,7 +14,7 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -24,6 +24,8 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
   provider:
+  flute_music_player: ^0.0.6
+  get_it:
 
 dev_dependencies:
   flutter_test:
@@ -44,6 +46,12 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
    - assets/top_banner.png
+   - assets/image1.png
+   - assets/image2.png
+   - assets/image3.png
+   - assets/image4.png
+   - assets/image6.jpg
+   - assets/image7.jpg
   #  - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
