@@ -1,14 +1,17 @@
 import 'package:flutter/material.dart';
-import 'package:musically/ui/shared/styles.dart';
-
 import 'SongBox.dart';
 
 class MusicList extends StatelessWidget {
   final List<String> list;
   final String heading;
   final Size size;
+  final List<String> image;
 
-  MusicList({@required this.list, @required this.heading, @required this.size});
+  MusicList(
+      {@required this.list,
+      @required this.heading,
+      @required this.size,
+      this.image});
 
   @override
   Widget build(BuildContext context) {
@@ -22,11 +25,10 @@ class MusicList extends StatelessWidget {
             padding: const EdgeInsets.only(left: 8.0),
             child: Text(
               heading,
-              style: Style.medium_bold_gothic(size.width),
             ),
           ),
           Container(
-            height: 250.0,
+            height: size.width/2,
             child: ListView.builder(
                 shrinkWrap: true,
                 scrollDirection: Axis.horizontal,
@@ -35,6 +37,7 @@ class MusicList extends StatelessWidget {
                   return MusicBox(
                     text: list[index],
                     size: size,
+                    image: image[index],
                   );
                 }),
           ),
