@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:musically/core/TabProvider.dart';
-import 'package:musically/ui/shared/styles.dart';
+import 'package:musically/core/enum/player_state.dart';
+import 'package:musically/core/viewmodel/TabProvider.dart';
+import 'package:musically/ui/locator.dart';
+import 'package:musically/ui/shared/color_styles.dart';
 import 'package:musically/ui/views/Music.dart';
-import 'package:musically/ui/views/Search.dart';
-import 'package:musically/ui/views/Settings.dart';
 import 'package:musically/ui/views/Home.dart';
-import 'package:provider/provider.dart';
+import 'package:musically/ui/widgets/base_view.dart';
 import '../SampleData.dart';
 
 class App extends StatefulWidget {
@@ -14,113 +14,147 @@ class App extends StatefulWidget {
 }
 
 class _AppState extends State<App> {
-  final List<Widget> _children = [Home(), Music(), Search(), Settings()];
+  final List<Widget> _children = [Home(), Music()];
 
   @override
   Widget build(BuildContext context) {
-    final currentIndex = Provider.of<TabProvider>(context);
-    final isPlaying = Provider.of<PlayProvider>(context);
     final width = MediaQuery.of(context).size.width;
     final Size size = MediaQuery.of(context).size;
     final height = MediaQuery.of(context).size.height;
-    return Scaffold(
-      body: _children[currentIndex.getCurrentTab > 1
-          ? currentIndex.getCurrentTab - 1
-          : currentIndex.getCurrentTab],
-      backgroundColor: Colors.white,
-      appBar: AppBar(
-        elevation: 0.0,
-        backgroundColor: Colors.white,
-        title: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: <Widget>[
-            Text(
-              'Musically',
-              style: Style.medium_bold_gothic(width),
-            ),
-            InkWell(
-              onTap: () {
-                Navigator.of(context).pushNamed('/setting');
-              },
-              child: ClipOval(
-                  child: CircleAvatar(
-                maxRadius: width / 18,
-                minRadius: width / 18,
-                child: Image.network(
-                  Data.profileImage,
-                  fit: BoxFit.cover,
-                ),
-              )),
-            ),
-          ],
-        ),
-      ),
-      bottomNavigationBar: BottomNavigationBar(
-        showUnselectedLabels: true,
-        elevation: 0.0,
-        type: BottomNavigationBarType.fixed,
-        onTap: (index) {
-          if (index != 2)
-            currentIndex.setCurrentTab(index);
-          else if (index == 2) {
-            isPlaying.play = !isPlaying.play;
-          }
-        },
-        backgroundColor: Colors.transparent,
-        selectedItemColor: Style.red,
-        unselectedItemColor: Style.gray,
-        items: [
-          BottomNavigationBarItem(
+    return BaseView<TabProvider>(
+        builder: (context, tab, _) => Scaffold(
+              body: _children[tab.getCurrentTab],
               backgroundColor: Colors.white,
-              icon: Icon(
-                Icons.home,
-                size: width / 16,
+              bottomNavigationBar: Container(
+                height: size.height / 6,
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.end,
+                  children: <Widget>[
+                    Column(
+                      children: <Widget>[
+                        Container(
+                          color: ColorStyle.gray,
+                          height: size.height / 19,
+                          child: Row(
+                            children: <Widget>[
+                              Container(
+                                  width: size.height / 18,
+                                  height: size.height / 18,
+                                  child: Image.asset(
+                                    Data.imageSet[1],
+                                    alignment: Alignment.centerLeft,
+                                    fit: BoxFit.cover,
+                                  )),
+                              Padding(
+                                padding: const EdgeInsets.only(left: 8.0),
+                                child: Column(
+                                  crossAxisAlignment: CrossAxisAlignment.start,
+                                  children: <Widget>[
+                                    Text(
+                                      'Bhut Hard',
+                                    ),
+                                    Text(
+                                      'Singles - Emiway',
+                                    ),
+                                  ],
+                                ),
+                              )
+                            ],
+                          ),
+                        ),
+                        SizedBox(
+                            width: size.width,
+                            height: 6.0,
+                            child: LinearProgressIndicator(
+                              value: .2,
+                              backgroundColor: ColorStyle.gray,
+                              valueColor:
+                                  AlwaysStoppedAnimation<Color>(ColorStyle.red),
+                            )),
+                      ],
+                    ),
+                    BaseView<PlayerProvider>(
+                      builder: (context, playprovider, _) =>
+                          BottomNavigationBar(
+                            showUnselectedLabels: true,
+                            elevation: 0.0,
+                            type: BottomNavigationBarType.fixed,
+                            onTap: (index) {
+                              if (index < 2)
+                                tab.setCurrentTab(index);
+                              else if (index == 2) {
+                                playprovider.playingState == PlayerState.playing
+                                    ? playprovider.playingState =
+                                        PlayerState.paused
+                                    : playprovider.playingState =
+                                        PlayerState.playing;
+                              } else if (index > 2) {
+                                if (index == 3) {
+                                  Navigator.of(context).pushNamed('/search');
+                                } else if (index == 4) {
+                                  Navigator.pushNamed(context, '/setting');
+                                }
+                              }
+                            },
+                            backgroundColor: Colors.transparent,
+                            selectedItemColor: ColorStyle.red,
+                            unselectedItemColor: ColorStyle.gray,
+                            items: [
+                              BottomNavigationBarItem(
+                                  backgroundColor: Colors.white,
+                                  icon: Icon(
+                                    Icons.home,
+                                    size: width / 16,
+                                  ),
+                                  title: Text(
+                                    'Home',
+                                  )),
+                              BottomNavigationBarItem(
+                                  icon: Icon(
+                                    Icons.library_music,
+                                    size: width / 16,
+                                  ),
+                                  title: Text(
+                                    'Music',
+                                  )),
+                              BottomNavigationBarItem(
+                                  icon: Icon(
+                                    playprovider.playingState ==
+                                            PlayerState.playing
+                                        ? Icons.play_circle_filled
+                                        : Icons.pause_circle_filled,
+                                    size: width / 8,
+                                    color: ColorStyle.red,
+                                  ),
+                                  title: Container(
+                                    height: 4.0,
+                                  )),
+                              BottomNavigationBarItem(
+                                  icon: Icon(
+                                    Icons.search,
+                                    size: width / 16,
+                                  ),
+                                  title: Text(
+                                    'Search',
+                                  )),
+                              BottomNavigationBarItem(
+                                  icon: Icon(
+                                    Icons.settings,
+                                    size: width / 16,
+                                  ),
+                                  title: Text(
+                                    'Setting',
+                                  ))
+                            ],
+                            currentIndex: tab.getCurrentTab,
+                          ),
+                    ),
+                  ],
+                ),
               ),
-              title: Text(
-                'Home',
-                style: Style.very_small_bold_gothic(width),
-              )),
-          BottomNavigationBarItem(
-              icon: Icon(
-                Icons.library_music,
-                size: width / 16,
-              ),
-              title: Text(
-                'Music',
-                style: Style.very_small_bold_gothic(width),
-              )),
-          BottomNavigationBarItem(
-              icon: Icon(
-                isPlaying.play == true
-                    ? Icons.play_circle_filled
-                    : Icons.pause_circle_filled,
-                size: width / 8,
-                color: Style.red,
-              ),
-              title: Container(
-                height: 4.0,
-              )),
-          BottomNavigationBarItem(
-              icon: Icon(
-                Icons.search,
-                size: width / 16,
-              ),
-              title: Text(
-                'Search',
-                style: Style.very_small_bold_gothic(width),
-              )),
-          BottomNavigationBarItem(
-              icon: Icon(
-                Icons.settings,
-                size: width / 16,
-              ),
-              title: Text(
-                'Setting',
-                style: Style.very_small_bold_gothic(width),
-              ))
-        ],
-        currentIndex: currentIndex.getCurrentTab,
-      ),
-    );
+            ));
   }
 }
+
+/*
+            */
