@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-class Style {
+class FontStyle {
   //define font style here
   static medium_bold_gothic(var size) => TextStyle(
       fontSize: size / 24.0, fontFamily: 'Gothic', color: Colors.black);
@@ -14,12 +14,6 @@ class Style {
   static medium_open_sans(var size) =>
       TextStyle(fontSize: size, fontFamily: 'Gothic', color: Colors.black);
 
-  static small_open_sans(var size) =>
-      TextStyle(fontSize: size, fontFamily: 'Gothic', color: Colors.black);
-
-  //define fonts color here
-  static final red = Color(0xfffc0506);
-  static final dark = Color(0xff2a2a2a);
-  static final gray = Color(0xffEBECED);
-  static final yellow = Color(0xfff49818);
+  static small_open_sans(var size) => TextStyle(
+      fontSize: size / 36, fontFamily: 'Gothic', color: Colors.black54);
 }
